+++ twitch/bot/src/commands/rewrite.rs
